requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
feedparser==6.0.10
beautifulsoup4==4.12.2
deep-translator==1.11.4
